+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Mapbox welcomes participation and contributions from everyone.
       * Removing default values for parameters
       * Making `bearing` and `pitch` parameters optional
       * Adding the `camera(for:camera:rect:)` variant
+  - `OrnamentOptions` should now be accessed via `MapView.ornaments.options`. `MapConfig.ornaments` has been removed. Updates can be applied directly to `OrnamentsManager.options`. Previously the map's ornament options were updated on `MapConfig.ornaments` with `MapView.update`. ([#310](https://github.com/mapbox/mapbox-maps-ios/pull/310)) 
+  - The `LogoView` class is now private. ([#310](https://github.com/mapbox/mapbox-maps-ios/pull/310))
+
+### Features ✨ and improvements 🏁
+
+- `OrnamentsManager` is now a public class and can be accessed via the `MapView`'s `ornaments` property.
 
 ## 10.0.0-beta.18.1 - April 28, 2021